frontend/Saloni/Speech_to_Speech_Saloni/requirements.txt
streamlit
azure-cognitiveservices-speech==1.35.0
python-dotenv==1.0.0
requests
pydub
